intermediate progress moving functions + state
